Move Google classes into their own package
