+++ ARCore/app/src/main/java/com/novoda/spikes/arcore/MainActivity.kt
@@ -9,7 +9,7 @@ import com.google.ar.core.exceptions.CameraNotAvailableException
 import com.novoda.spikes.arcore.helper.ARCoreDependenciesHelper
 import com.novoda.spikes.arcore.helper.CameraPermissionHelper
 import com.novoda.spikes.arcore.rendering.NovodaSurfaceViewRenderer
-import com.novoda.spikes.arcore.helper.TapHelper
+import com.novoda.spikes.arcore.google.helper.TapHelper
 import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
