+++ ARCore/app/src/main/java/com/novoda/spikes/arcore/google/helper/TapHelper.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.novoda.spikes.arcore.helper;
+package com.novoda.spikes.arcore.google.helper;
 
 import android.content.Context;
 import android.view.GestureDetector;
