+++ ARCore/app/src/main/java/com/novoda/spikes/arcore/google/rendering/PlaneRenderer.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.novoda.spikes.arcore.rendering;
+package com.novoda.spikes.arcore.google.rendering;
 
 import android.content.Context;
 import android.graphics.Bitmap;
