+++ ARCore/app/src/main/java/com/novoda/spikes/arcore/visualiser/PlanesVisualiser.kt
@@ -5,7 +5,8 @@ import com.google.ar.core.Camera
 import com.google.ar.core.Plane
 import com.google.ar.core.Session
 
-import com.novoda.spikes.arcore.rendering.PlaneRenderer
+import com.novoda.spikes.arcore.google.rendering.PlaneRenderer
+import com.novoda.spikes.arcore.rendering.ARCoreDataModel
 
 class PlanesVisualiser(private val context: Context) {
     private val planeRenderer = PlaneRenderer()
